Convert parameter to long
git-svn-id: file:///home/svn/DVA2006@1993 1e6d2d28-e1fb-0310-b8ab-81908d8c6b6a

+++ src/java/org/deltava/util/CalendarUtils.java
@@ -18,9 +18,9 @@ public class CalendarUtils {
 	 * @param days the number of days to adjust
 	 * @return the adjusted date/time
 	 */
-	public static Date adjust(Date dt, int days) {
+	public static Date adjust(Date dt, long days) {
 		Calendar cld = getInstance(dt);
-		cld.add(Calendar.DATE, days);
+		cld.add(Calendar.DATE, (int) days);
 		return cld.getTime();
 	}
